+++ test/test_fitprior.jl
@@ -3,7 +3,7 @@
 @test_nothrow fitprior(distM, "k-medoids", true; verbose = false) # kmedoids with distances, Julia
 @test_logs (:warn,) global params = fitprior(pnts, "k-means", false; verbose = false, Kmin = 1, Kmax = 1) # check edge case Kmax = 2
 @test params.K_initial == 1
-@test_logs (:warn,) global params = fitprior(pnts, "k-means", false; verbose = false, Kmin = N, Kmax = N) # check edge case Kmin = N-1
+@test_logs (:warn,) global params = fitprior(pnts, "k-means", false; verbose = false, Kmin = N, Kmax = N) # check edge case Kmin = N
 @test params.K_initial == N
 @test_throws ArgumentError fitprior(distM, "hierarchical", true; verbose = false) # check that only 2 methods allowed
 @test_throws ArgumentError fitprior(pnts, "k-means", true; verbose = false) # check that dist is disregarded when pnts is supplied
@@ -19,5 +19,5 @@
 @test_nothrow fitprior2(distM, "k-medoids", true; verbose = false) # kmedoids with distances, Julia
 @test_logs (:warn,) global params = fitprior2(pnts, "k-means", false; verbose = false, Kmin = 1, Kmax = 1) # check edge case Kmax = 2
 @test params.K_initial == 1
-@test_logs (:warn,) global params = fitprior2(pnts, "k-means", false; verbose = false, Kmin = N, Kmax = N) # check edge case Kmin = N-1
+@test_logs (:warn,) global params = fitprior2(pnts, "k-means", false; verbose = false, Kmin = N, Kmax = N) # check edge case Kmin = N
 @test params.K_initial == N